New Footer component w/ material UI
