+++ src/pages/legacy.tsx
@@ -6,7 +6,7 @@ import Container from "react-bootstrap/Container"
 import Col from "react-bootstrap/Col"
 import Row from "react-bootstrap/Row"
 
-import DefaultFooter from "../components/footer"
+import DefaultFooter from "../components/Footer"
 import SEO from "../components/SEO"
 import { availableIcons, Icon } from "../components/icons"
 import stacks from "../meta/stacks"
